Added DAE1 Framework Physics files and added Command pattern
I added the physics related files from the first framework from the DAE instead of Box2D, which is harder/longer to implement and work with. And I started to create a Box2D component. But it is not as easy to find out where and how exactly to check the physic.